requirements.txt
opencv-python==4.11.0.86
rembg==2.0.65
pillow==10.4.0
onnxruntime==1.21.0